refactor: [#453] reorganize console mods
```console
src/console/
├── commands
│   ├── mod.rs
│   └── tracker_statistics_importer
│       ├── app.rs
│       └── mod.rs
├── cronjobs
│   ├── mod.rs
│   └── tracker_statistics_importer.rs
└── mod.rs
```

- We will create a new command to seed the Index with random torrents.
- This new strcuture is similar to what we have in the Tracker.
