+++ src/console/commands/tracker_statistics_importer/mod.rs
@@ -1,0 +1,1 @@
+pub mod app;
